reduce active meeting subscription to minimum
